+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { Route, Routes } from "react-router-dom";
 import { useState, useEffect } from "react";
 import styled from "styled-components";
-import Home from "./components/Home";
-import Footer from "./components/Footer";
+import Home from "./components/firstPage/Home";
+import Footer from "./components/firstPage/Footer";
 import Navbar from "./components/Navbar";
 import AboutUs from "./components/aboutUs/AboutUs";
 import LoginForm from "./components/LoginForm";
