+++ src/components/firstPage/hero/Hero.tsx
