+++ src/components/firstPage/hero/HeroLeft.tsx
