+++ src/components/firstPage/hero/HeroRight.tsx
